Trait findshlibs::Segment

pub trait Segment: Sized + Debug {
    type SharedLibrary: SharedLibrary<Segment = Self>;

    // Required methods
    fn name(&self) -> &str;
    fn stated_virtual_memory_address(&self) -> Svma;
    fn len(&self) -> usize;

    // Provided methods
    fn is_code(&self) -> bool { ... }
    fn is_load(&self) -> bool { ... }
    fn actual_virtual_memory_address(&self, shlib: &Self::SharedLibrary) -> Avma { ... }
    fn contains_svma(&self, address: Svma) -> bool { ... }
    fn contains_avma(&self, shlib: &Self::SharedLibrary, address: Avma) -> bool { ... }
}

A mapped segment in a shared library.

Required Associated Types

type SharedLibrary: SharedLibrary<Segment = Self>

The associated shared library type for this segment.

Required Methods

fn name(&self) -> &str

Get this segment’s name.

fn stated_virtual_memory_address(&self) -> Svma

Get this segment’s stated virtual address of this segment.

This is the virtual memory address without the bias applied. See the module documentation for details.

fn len(&self) -> usize

Get the length of this segment in memory (in bytes).

Provided Methods

fn is_code(&self) -> bool

Returns true if this is a code segment.

fn is_load(&self) -> bool

Returns true if this is a segment loaded into memory.

fn actual_virtual_memory_address(&self, shlib: &Self::SharedLibrary) -> Avma

Get this segment’s actual virtual memory address.

This is the virtual memory address with the bias applied. See the module documentation for details.

fn contains_svma(&self, address: Svma) -> bool

Does this segment contain the given address?

fn contains_avma(&self, shlib: &Self::SharedLibrary, address: Avma) -> bool

Does this segment contain the given address?

Object Safety

This trait is not object safe.