Trait findshlibs::Segment

pub trait Segment: Sized + Debug {
    type SharedLibrary: SharedLibrary<Segment = Self>;
    fn name(&self) -> &str;

    fn stated_virtual_memory_address(&self) -> Svma;

    fn len(&self) -> usize;

    fn is_code(&self) -> bool { ... }

    fn actual_virtual_memory_address(&self, shlib: &Self::SharedLibrary) -> Avma { ... }

    fn contains_svma(&self, address: Svma) -> bool { ... }

    fn contains_avma(&self, shlib: &Self::SharedLibrary, address: Avma) -> bool { ... }
}

A mapped segment in a shared library.

Associated Types

type SharedLibrary: SharedLibrary<Segment = Self>

The associated shared library type for this segment.

Loading content...
Required methods

fn name(&self) -> &str

Get this segment's name.

fn stated_virtual_memory_address(&self) -> Svma

Get this segment's stated virtual address of this segment.

This is the virtual memory address without the bias applied. See the module documentation for details.

fn len(&self) -> usize

Get the length of this segment in memory (in bytes).

Loading content...
Provided methods

fn is_code(&self) -> bool

Returns true if this is a code segment.

fn actual_virtual_memory_address(&self, shlib: &Self::SharedLibrary) -> Avma

Get this segment's actual virtual memory address.

This is the virtual memory address with the bias applied. See the module documentation for details.

fn contains_svma(&self, address: Svma) -> bool

Does this segment contain the given address?

fn contains_avma(&self, shlib: &Self::SharedLibrary, address: Avma) -> bool

Does this segment contain the given address?

Loading content...