Trait findshlibs::SharedLibrary

pub trait SharedLibrary: Sized + Debug {
    type Segment: Segment<SharedLibrary = Self>;
    type SegmentIter: Debug + Iterator<Item = Self::Segment>;
    fn name(&self) -> &OsStr;

    fn id(&self) -> Option<SharedLibraryId>;

    fn segments(&self) -> Self::SegmentIter;

    fn virtual_memory_bias(&self) -> Bias;

    fn each<F, C>(f: F)
    where
        F: FnMut(&Self) -> C,
        C: Into<IterationControl>;

    fn avma_to_svma(&self, address: Avma) -> Svma { ... }
}

A trait representing a shared library that is loaded in this process.

Associated Types

type Segment: Segment<SharedLibrary = Self>

The associated segment type for this shared library.

type SegmentIter: Debug + Iterator<Item = Self::Segment>

An iterator over a shared library's segments.

Loading content...
Required methods

fn name(&self) -> &OsStr

Get the name of this shared library.

fn id(&self) -> Option<SharedLibraryId>

Get the debug-id of this shared library if available.

fn segments(&self) -> Self::SegmentIter

Iterate over this shared library's segments.

fn virtual_memory_bias(&self) -> Bias

Get the bias of this shared library.

See the module documentation for details.

fn each<F, C>(f: F) where
F: FnMut(&Self) -> C,
C: Into<IterationControl>,

Find all shared libraries in this process and invoke f with each one.

Loading content...
Provided methods

fn avma_to_svma(&self, address: Avma) -> Svma

Given an AVMA within this shared library, convert it back to an SVMA by removing this shared library's bias.

Loading content...